Allow Term.readInput to be interrupted by signals
This allows a well-written program to update immediately on SIGWINCH. Well written programs should
not do any redraw logic in a signal handler, because a signal handler can be invoked at any point
in the programs execution. Programs should instead set an atomic flag that is checked at the next
redraw call. Interrupting the read call allows programs to immediately proceed to the next
redraw call, and update immediately. Old behaviour can be restored by setting the SA.RESTART flag
on creation of a signal handler.