Fix utf-8 codepoints being cut off in readInput
Previously utf-8 codepoints could be cut off if there was not enough room at the end of the given
buffer for the entire codepoint. This would lead to an out-of-bounds indexing when parsing the
input later. This commit detects if there is a cut-off codepoint at the end of the buffer, and if
there is, stores it in an internal buffer to be used at the next call to readInput.