Store canvas text per line
Canvas' will mostly be used for complex drawing operations where there
are lots of sources drawing to the canvas at arbitrary locations, such
as a hierarchy of graphical widgets. This requires good random write
performance.

The old design used a single buffer for all canvas text, which reduced
the amount of allocations and work required when resizing. But it raised
performance concerns when doing lots of random writes on the canvas, as
the text buffer would have to do lots of large memcopies on every write.

Changing to a buffer-per-line approach should reduce the amount of bytes
being copied around considerably and speed up writing, which is the most
common operation. It probably has some impact on resize performance, but
it wouldn't be that much.