Fix 24bit bg colors, support truecolor checks
Fixes a bug where 24bit background colors were accidentally being output
as foreground colors.

Truecolor can now be force enabled or disabled when initializing a Term
instance. The `dump` function for styles has been modified to accept an
option for truecolor support.

Some more logging has been added for investigating the number of bytes
sent to the terminal on an update. This can be disabled by restricting
the scope of `shovel` and `shovel_perf` in your std_options or in a
custom log function.