Overhaul terminfo, introduce Screen API
Overhauls terminfo handling. All places that take a terminfo no longer
take an optional pointer, they now instead take a normal pointer to a
TermInfo instance. To not use TermInfo, simply pass an empty (default
initialised) TermInfo instance. APIs that queried TermInfo (like the
functions in Term.RenderContext) no longer use arbitrary fallback
sequences. When a requested capability is not found, they now simply
do nothing.

Adds a way to specify fallback terminal capabilities. There are two ways
to specify fallback definitions: a list of TERM variables, from which
the first valid terminfo found will be used, and a custom terminfo
source. Shovel now ships two terminfo definitions that can be specified
via decl literals, those being `dumb` and `xterm-256color`. There are
three fallback modes that determine how the fallback terminfo will be
used. `last_resort` will only use the fallback if a terminfo file cannot
be loaded for the current TERM. `always` will ignore the current TERM
and only use the specified fallback terminfo. `merge` will use the
current TERM, and then use the fallback for any sequences not defined by
TERM.

24-bit colour (truecolour) support has been improved. Term.init now
checks for setrgbf/setrgbb capabilities, the Tc flag, the RGB flag, and
the COLORTERM environment variable. Dumping a 24-bit colour will now
use the setrgb sequences if found, hardcoded sequences if the Tc flag
or COLORTERM variables are defined, and setaf/setab if the RGB flag is
defined. If none of these are defined, 24-bit colours are approximated
using the 256 colour palette. Note that terminfo definitions that
use setaf/setab for 24-bit colours are broken by design - this is not
something shovel can fix.

Adds a `diff` option to Style.dump. This can be used to only dump the
difference between two styles.

Corrects handling of `.none` foreground/background colours. The previous
implementation treated them as not doing anything, whereas they now
reset the foreground/background colour to the terminal's default.

Adds an example for truecolor capabilities.

TermInfo.parse now takes a reader instead of a buffer.

Introduces a new API, shovel.Screen. Screens are buffers that represent
the screen of a terminal. They can be resized and written to as if they
were a terminal screen, and can write their contents to a given writer.
The primary motivation of this new API is to make it easy to write only
the minimum necessary bytes to update the terminal. The previous
RenderContext API would happily write redundant bytes to the terminal,
so users would have to make sure to avoid writing over the same part
of the screen multiple times to avoid flickering. Users would also have
to manually implement caching to avoid sending too many redundant escape
sequences to the terminal.

The Screen API correctly handles unicode grapheme clusters, with a fast
path for ASCII text. It has a Writer that allows for drawing at specific
rectangles on the screen, with truncating or wrapping behaviour when
reaching the end of the rectangle, depending on the current wrapping
mode.

Introduces double buffering via the shovel.Screen.DoubleBuffer API. It
stores two screens, a read screen and a write screen. The read screen
contains the current state of the physical terminal's screen. The write
screen is updated manually to draw to the screen. The double buffer
has a dump method that writes the minimal amount of bytes required to
transform the read screen into the write screen, and then copies the
write screen's contents to the read screen so they are identical. Lines
that have not changed between the screens are not outputted.