Screen: more wide character fixes
Now correctly erases wide characters when you write over the right-side
column of an existing wide character.

Fixes incorrectly wrapping the cursor when writing zero width
characters.

Now manually moves the cursor after dumping a wide character. This is
needed because wide characters are not always treated as wide by every
terminal emulator. For example, the Linux TTY prints emoji codepoints
as a single column, while it correctly treats CJK characters as two
columns. There is also differences in the width of graphemes like
3-em-dash between terminal emulators. Querying the width of every
character is not reliable and not supported on every terminal emulator,
so the only sane way to handle this is to manually position the cursor
after dumping a wide character. For terminals without the ability to
address the cursor, a single space is always written after a wide
character. Terminals that can't address the cursor almost certainly
won't support unicode anyway, so this is fine.

writeTruncating now correctly pads the output if the first truncated
character was a wide character. Previously it would leave a space of one
because it was replacing the wide character with a single-width triple
elipses character.

Term.uncook no longer initializes unicode data, and as such no longer
takes an allocator. Programs should initialize Shovel's unicode data
themselves by calling shovel.initUnicodeData and
shovel.deinitUnicodeData.